+++ Lib/gftools/scripts/push_status.py
@@ -31,35 +31,8 @@ gftools push-status /path/to/google/fonts/repo --lint
 """
 import argparse
 from pathlib import Path
-from gftools.push2 import push_report, PushItems
-
-
-def lint_server_files(fp):
-    template = "{}: Following paths are not valid:\n{}\n\n"
-    footnote = (
-        "lang and axisregistry dir paths need to be transformed.\n"
-        "See https://github.com/googlefonts/gftools/issues/603"
-    )
-
-    prod_path = fp / "to_production.txt"
-    production_file = PushItems.from_server_file(prod_path, "In Sandbox")
-    prod_missing = "\n".join(map(str, production_file.missing_paths()))
-    prod_msg = template.format("to_production.txt", prod_missing)
-
-    sandbox_path = fp / "to_sandbox.txt"
-    sandbox_file = PushItems.from_server_file(sandbox_path, "In Dev / PR Merged")
-    sandbox_missing = "\n".join(map(str, sandbox_file.missing_paths()))
-    sandbox_msg = template.format("to_sandbox.txt", sandbox_missing)
-
-    if prod_missing and sandbox_missing:
-        raise ValueError(prod_msg + sandbox_msg + footnote)
-    elif prod_missing:
-        raise ValueError(prod_msg + footnote)
-    elif sandbox_missing:
-        raise ValueError(sandbox_msg + footnote)
-    else:
-        print("Server files have valid paths")
-
+from gftools.push2 import push_report, lint_server_files
+import os
 
 def main(args=None):
     parser = argparse.ArgumentParser()
@@ -68,6 +41,10 @@ def main(args=None):
         "--lint", action="store_true", help="Check server files have correct paths"
     )
     args = parser.parse_args(args)
+
+    if "ofl" not in os.listdir("."):
+        raise ValueError("tool must be run from a google/fonts dir")
+
     if args.lint:
         lint_server_files(args.path)
     else:
